Create Swatch component to lighten up ManiMaker
